#1301 clean up and logging in once at start up
